Blacklist Caskaydia nerd fonts (causes issues)
Caskaydia Nerd Fonts are automatically installed with `nerd-fonts-git`
AUR package, however their presence seems to be causing issues with
Material Symbols font. This then affects the eww bar, making the icons
small.

This is a pretty odd issue, but after some debugging, removing or
blacklisting this font does resolve the issue. It seems that it's some
font priority issue, however the eww bar clearly specifies which font to
use, and the icons are correct, it's just the size that's weird (too
small).

This is just a hacky workaround, but it's probably a permanent one,
unless I figure out what exactly is the cause of this and resolve it
diferently, which I'm probably too lazy to get to.