Only add ~/.local/bin to PATH if it exists
Without existence check, the `find` command prints a fail message to
stderr, and leaves stdout empty. This means only the prefix `:` is
appended to the `$PATH` variable.

Leaving a trailing `:` in `$PATH` is however something we really don't
want to be doing, as it makes the current directory get treated as a
part of `$PATH`.

While the `~/.local/bin` directory is generally going to exist, during
installation and in some other edge cases, it might not yet be there,
and we need to account for that.